+++ src/lib/components/Tooltip.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import { Tooltip } from "bits-ui";
-	import { fly } from "svelte/transition";
+	import { fade } from "svelte/transition";
 
 	export let text: string = ""
 	export let side: "top" | "right" | "bottom" | "left" = "top";
 	export let sideOffset: number = 0;
-	export let openDelay: number = 150;
+	export let openDelay: number = 100;
 
 	let triggerClass: string = "";
 
@@ -18,8 +18,8 @@
 	</Tooltip.Trigger>
 	<Tooltip.Content
 		class="rounded-lg bg-gray-800 border border-gray-600 text-slate-300 px-4 py-2 shadow-lg max-w-[35rem]"
-		transition={fly}
-		transitionConfig={{ duration: 100 }}
+		transition={fade}
+		transitionConfig={{ duration: 50 }}
 		{sideOffset}
 		{side}
 	>