+++ src/lib/config/FlagsConfig.svelte
@@ -28,6 +28,5 @@
 	bind:selected
 	{onSelectedChange}
 	multiple={true}
-	size='sm'
 />
 <ResetConfigButton {entryId} {onReset} />